requirements.txt
PyYAML
Pillow
opencv-python
matplotlib
ipykernel